Use IMAP UID to identify messages in the database.

+++ app/src/main/java/de/fuchsch/email/database/entity/MessageEntity.kt
@@ -16,6 +16,6 @@ data class MessageEntity (
     val message: String,
     val sender: String,
     val recipients: List<String>,
-    val messageNumber: Int,
+    val messageNumber: Long,
     val folder: String
 )
